Split out TablePropertiesAsyncCommitsTest
